+++ src/test/resources/unmapped/C-CDA_R2-1_CCD-ProcedureActivityProcedure-unmapped.xml
@@ -44,6 +44,8 @@
                   </specimenPlayingEntity>
                </specimenRole>
             </specimen>
+            <performer>
+							</performer>
             <participant typeCode="DEV">
                <participantRole classCode="MANU">
 									<!-- ** Product instance ** -->
